ADD: Rest Interfaces ADD: Documentation for interfaces and non trivial methods on entities
