+++ projects/src/main/java/com/sbs/projects/models/Sprint.java
@@ -2,7 +2,14 @@ package com.sbs.projects.models;
 
 import com.sbs.contracts.dto.SprintDTO;
 
-import javax.persistence.*;
+import javax.persistence.Basic;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import java.time.LocalDate;
 
 @Entity
@@ -74,6 +81,11 @@ public class Sprint {
         this.project = project;
     }
 
+    /**
+     * Transforms the Sprint Entity into a SprintDTO object.
+     *
+     * @return the SprintDTO corresponded to the Entity
+     */
     public SprintDTO toDTO() {
         return new SprintDTO.Builder()
                 .withId(this.id)
@@ -83,6 +95,11 @@ public class Sprint {
                 .build();
     }
 
+    /**
+     * Mocks the relationship betweeen Sprints and Project with the given <code>projectId</code>
+     *
+     * @param projectId the Project ID to mock the relationship
+     */
     public void setIsolatedProject(Long projectId) {
         Project project = new Project();
         project.setId(projectId);
