+++ userStories/src/main/java/com/sbs/userStories/services/LabelService.java
@@ -4,6 +4,11 @@ import com.sbs.contracts.dto.LabelDTO;
 
 public interface LabelService {
 
+    /**
+     * Gets all the labels saved
+     *
+     * @return all the labelDTO objects saved
+     */
     Iterable<LabelDTO> getAll();
 
 }
